core: Refactor listeners; use SO_REUSEPORT on Unix
Just an experiment for now
